End of chapter2

+++ run.bash
@@ -2,27 +2,47 @@
 set -eou pipefail
 
 
-results_base_path="$CSC500_ROOT_PATH/csc500-past-runs/chapter2/withTipoffRegularConv"
+# results_base_path="$CSC500_ROOT_PATH/csc500-past-runs/chapter2/withTipoffRegularConv"
+results_base_path="$CSC500_ROOT_PATH/csc500-past-runs/chapter2/doesTipoffFuckUs"
 
 patience=10
 
 ALL_SERIAL_NUMBERS='["3123D52","3123D65","3123D79","3123D80","3123D54","3123D70","3123D7B","3123D89","3123D58","3123D76","3123D7D","3123EFE","3123D64","3123D78","3123D7E","3124E4A"]'
 ALL_DISTANCES="[14,2,44,62,20,32,50,8,26,38,56]"
 
-for seed in 2293 15474 1924 25792 6107 6031 1710 4253 8134 5133; do
+
+# The Hail Merry tipoff settings
+# for seed in 2293 15474 1924 25792 6107 6031 1710 4253 8134 5133; do
+# for batch_size in 64 128 256 512; do
+# for epochs in 10000; do
+# #for learning_rate in 0.0001; do
+# for learning_rate in 0.00001 0.001; do
+# for source_distance in $ALL_DISTANCES; do
+# for target_distance in $ALL_DISTANCES; do
+# for alpha in 0.001; do
+# for window_stride in 50; do
+# for window_length in 256; do
+# for num_examples_per_device in 50000; do
+# # for num_examples_per_device in 500; do
+# for desired_runs in "[1]"; do
+# for desired_serial_numbers in "$ALL_SERIAL_NUMBERS"; do
+# for tipoff in "true"; do
+
+
+for seed in 2293 15474 1924; do
 for batch_size in 256; do
-for epochs in 10000; do
+for epochs in 100; do
 for learning_rate in 0.0001; do
-for source_distance in $ALL_DISTANCES; do
-for target_distance in $ALL_DISTANCES; do
+for source_distance in "[14]"; do
+for target_distance in "[14]"; do
 for alpha in 0.001; do
 for window_stride in 50; do
 for window_length in 256; do
-for num_examples_per_device in 200000; do
+for num_examples_per_device in 10000; do
 for desired_runs in "[1]"; do
 for desired_serial_numbers in "$ALL_SERIAL_NUMBERS"; do
-for tipoff in "true"; do
-    experiment_name=name:cnnOnly-nllLoss
+for tipoff in "true" "false"; do
+    experiment_name=name:cnnOnly-nllLoss-fakeTipoff
     experiment_name=${experiment_name}_tipoff:$tipoff
     experiment_name=${experiment_name}_seed:${seed}
     experiment_name=${experiment_name}_learningRate:${learning_rate}
@@ -51,6 +71,13 @@ for tipoff in "true"; do
     else
         experiment_name=${experiment_name}_targetDistance:$target_distance
     fi
+
+    # Check if experiment already exists, if it does we skip
+    if [ -d "$results_base_path/$experiment_name" ]; then
+        echo "$results_base_path/$experiment_name already exists, skipping"
+        continue
+    fi
+    
 
     rm -rf results
     mkdir results
